+++ schemadiff/__init__.py
@@ -4,7 +4,7 @@ from graphql import GraphQLSchema as GQLSchema, is_schema
 
 from schemadiff.changes import Change
 from schemadiff.diff.schema import Schema
-from schemadiff.graphql_schema import Schema
+from schemadiff.schema_loader import Schema
 from schemadiff.formatting import print_diff, format_diff
