+++ tests/test_formatting.py
@@ -1,7 +1,7 @@
 import json
 
 from schemadiff.diff.schema import Schema
-from schemadiff.graphql_schema import Schema
+from schemadiff.schema_loader import Schema
 from schemadiff.formatting import print_diff, print_json
 from tests.test_schema_loading import TESTS_DATA
